Added multiple commits in main_md_table.md - issue #113.
In order to add multiple commits I had to:
- Redefine clear_file() (added new parameter, modified the heading and
removed the base table)
- Created a new function "generate_markdown_header()" (that has the base
table from clear_file() in which I write the repository_title and
title_table_formatation + the base_table
- Modified write_main_md_table() function to take as an argument the
repository url
- Modified extract_json() function, added a new parameter (commits_per_repo)
which by default it's 5 and represents the number of commits to be used in
the main markdown file - this can be modified to make the function write
as many commits per repository as we want. extract_json() has pretty much
retained it's function but redesigned and we now generate the markdown header
for each repository, and after that we add the commits.
All of the commits and markdown headers are appended untill the table
generation is completed.